move get_parameter_from_url_or_session to evaluation app
