+++ evap/staff/tools.py
@@ -19,16 +19,6 @@ from evap.grades.models import GradeDocument
 from evap.results.tools import collect_results
 
 
-def get_parameter_from_url_or_session(request, parameter):
-    result = request.GET.get(parameter, None)
-    if result is None:  # if no parameter is given take session value
-        result = request.session.get(parameter, False)  # defaults to False if no session value exists
-    else:
-        result = {'true': True, 'false': False}.get(result.lower())  # convert parameter to boolean
-    request.session[parameter] = result  # store value for session
-    return result
-
-
 def forward_messages(request, success_messages, warnings):
     for message in success_messages:
         messages.success(request, message)
